unifiy proxy logic and refactor the basics
