+++ Cargo.toml
@@ -21,7 +21,7 @@ once_cell = "1.12.0"
 url = { version = "2.2.2", features = ["serde"] }
 redis = { version = "0.21.5", default-features = false, features = ["tokio-comp"] }
 serde = "1.0.137"
-serde_json = "1.0.81"
+serde_json = { version = "1.0.81", features = ["preserve_order", "raw_value"] }
 sha3 = "0.9"
 simple_logger = "2.1.0"
 tokio = { version = "1.12.0", default-features = false, features = ["macros", "rt-multi-thread", "sync", "time"] }
