+++ src/proxy/websocket.rs
@@ -9,11 +9,7 @@ use tokio_tungstenite::{
 };
 
 use crate::{
-    ctx::AppConfig,
-    http::response_by_status,
-    log_format,
-    proxy::{validation_middleware_quicknode, QuicknodeSocketPayload},
-    GenericResult,
+    ctx::AppConfig, http::response_by_status, log_format, rpc::RpcSocketPayload, GenericResult,
 };
 
 pub(crate) fn should_upgrade_to_socket_conn(req: &Request<Body>) -> bool {
@@ -137,7 +133,7 @@ pub(crate) async fn socket_handler(
                                             match msg {
                                                 Some(Ok(msg)) => {
                                                     if let Message::Text(msg) = msg {
-                                                         let socket_payload: QuicknodeSocketPayload = match serde_json::from_str(&msg) {
+                                                         let socket_payload: RpcSocketPayload = match serde_json::from_str(&msg) {
                                                              Ok(t) => t,
                                                              Err(e) => {
                                                                  if let Err(e) = inbound_socket.send(format!("Invalid payload. {e}").into()).await {
@@ -173,8 +169,7 @@ pub(crate) async fn socket_handler(
                                                              continue;
                                                         }
 
-                                                        // TODO add general validation_middleware support (if have new features which support websocket)
-                                                        match validation_middleware_quicknode(
+                                                        match crate::proxy::http::post::validation_middleware(
                                                             &cfg,
                                                             &signed_message,
                                                             &proxy_route,
